[WS-147] Adds a new argument to "/setwarp" and "/delwarp" to directly overwrite/delete warps * Moved SimpleWarps from Warps to an own directory
