+++ src/de/codingair/warpsystem/spigot/features/simplewarps/utils/actions/types/PayAction.java
@@ -1,8 +1,8 @@
-package de.codingair.warpsystem.spigot.features.warps.simplewarps.utils.actions.types;
+package de.codingair.warpsystem.spigot.features.simplewarps.utils.actions.types;
 
 import de.codingair.warpsystem.spigot.base.utils.money.Adapter;
 import de.codingair.warpsystem.spigot.base.utils.money.MoneyAdapterType;
-import de.codingair.warpsystem.spigot.features.warps.simplewarps.utils.actions.Action;
+import de.codingair.warpsystem.spigot.features.simplewarps.utils.actions.Action;
 import org.bukkit.entity.Player;
 
 public class PayAction implements Action {
